TeachPendant/mcp_servers/program_repository/pyproject.toml
[project]
name = "program_repository"
version = "0.1.0"
description = "MCP server for TeachPendant customer program queries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=0.9.0",
  "pyyaml>=6.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["program_repository"]
